fetcher/pyproject.toml
[tool.poetry]
name = "fetcher"
version = "0.1.0"
description = ""
authors = ["Joshua Tompkins <josh@joshtompkins.com>"]

[tool.poetry.dependencies]
python = "^3.8"
readability-lxml = "^0.8.1"
lxml = "^4.5.2"
requests = "^2.24.0"
kermes-infra = {path = "../libs/kermes-infra"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = {version = "^20.8b1", allow-prereleases = true}
flake8-pytest = "^1.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
